eRPW: Workflow changes on develop
