+++ docroot/profiles/unicef/modules/custom/erpw_location/erpw_location.services.yml
@@ -4,6 +4,6 @@ services:
     arguments: ['@entity_type.manager', '@language_manager','@database']
   erpw_location.location_cookie:
     class: Drupal\erpw_location\LocationCookie
-    arguments: ['@request_stack', '@current_route_match', '@database', '@entity_type.manager']
+    arguments: ['@request_stack']
     tags:
       - { name: 'event_subscriber' }
