+++ docroot/profiles/unicef/modules/custom/erpw_location/modules/erpw_pathway/erpw_pathway.services.yml
@@ -16,3 +16,12 @@ services:
       - "@entity_type.manager"
       - "@erpw_location.location_services"
       - "@language_manager"
+  erpw_pathway.workflow_moderation:
+    class: Drupal\erpw_pathway\EventSubscriber\RemoveRejectedModeration
+    arguments:
+      - '@request_stack'
+      - '@current_route_match'
+      - '@database'
+      - '@entity_type.manager'
+    tags:
+      - { name: 'event_subscriber' }
