+++ docroot/profiles/unicef/modules/custom/erpw_location/src/LocationCookie.php
@@ -2,9 +2,6 @@
 
 namespace Drupal\erpw_location;
 
-use Drupal\Core\Database\Connection;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
-use Drupal\Core\Routing\RouteMatchInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Cookie;
 use Symfony\Component\HttpFoundation\RequestStack;
@@ -19,32 +16,11 @@ use Symfony\Component\HttpKernel\KernelEvents;
 class LocationCookie implements EventSubscriberInterface {
 
   /**
-   * Database instance.
-   *
-   * @var \Drupal\Core\Database\Connection
-   */
-  protected $connection;
-
-  /**
-   * A entityTypeManager instance.
-   *
-   * @var Drupal\Core\Entity\EntityTypeManagerInterface
-   */
-  protected $entityTypeManager;
-
-  /**
    * Current request.
    *
    * @var \Symfony\Component\HttpFoundation\Request|null
    */
   protected $request;
-
-  /**
-   * The current route match.
-   *
-   * @var \Drupal\Core\Routing\RouteMatchInterface
-   */
-  protected $routeMatch;
 
   /**
    * Name of the cookie this service will manage.
@@ -79,21 +55,9 @@ class LocationCookie implements EventSubscriberInterface {
    *
    * @param \Symfony\Component\HttpFoundation\RequestStack $request_stack
    *   Request stack service.
-   * @param \Drupal\Core\Routing\RouteMatchInterface $route_match
-   *   Route match service.
-   * @param \Drupal\Core\Database\Connection $connection
-   *   Database service.
-   * @param Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
-   *   EntityManager object.
    */
-  public function __construct(RequestStack $request_stack,
-    RouteMatchInterface $route_match,
-    Connection $connection,
-    EntityTypeManagerInterface $entity_type_manager) {
+  public function __construct(RequestStack $request_stack) {
     $this->request = $request_stack->getCurrentRequest();
-    $this->routeMatch = $route_match;
-    $this->connection = $connection;
-    $this->entityTypeManager = $entity_type_manager;
   }
 
   /**
@@ -181,10 +145,6 @@ class LocationCookie implements EventSubscriberInterface {
    */
   public function onResponse(ResponseEvent $event) {
     $response = $event->getResponse();
-    $route_name = $this->routeMatch->getRouteName();
-    if ($route_name == 'view.manage_service.page_1' || $route_name == 'view.service_list.page_1') {
-      $this->removeModeratedContent();
-    }
     if ($this->getShouldUpdateCookie()) {
       $domain_current_url = explode(".", $this->request->server->get('SERVER_NAME'));
       $domain_slice = array_slice($domain_current_url, -2);
@@ -198,34 +158,6 @@ class LocationCookie implements EventSubscriberInterface {
     if ($this->getShouldDeleteCookie()) {
       $response->headers->clearCookie($this->getCookieName());
     }
-  }
-
-  /**
-   * Remove the moderated content of the reject.
-   */
-  public function removeModeratedContent() {
-    $query = $this->connection->select('content_moderation_state_field_revision', 'cm');
-    $query->fields('cm', ['content_entity_id'])
-      ->condition('moderation_state', 'reject', '=')
-      ->condition('workflow', 'service_editorial');
-    $query = $query->execute();
-    $nids = $query->fetchCol();
-    if (!empty($nids)) {
-      $query = $this->connection->select('content_moderation_state_field_revision', 'cm');
-      $query->leftjoin('node_revision', 'n', 'n.vid = cm.content_entity_revision_id');
-      $query->fields('cm', ['content_entity_revision_id', 'moderation_state']);
-      $query->fields('n', ['vid'])
-        ->condition('cm.content_entity_id', $nids, 'IN')
-        ->condition('n.revision_default', '0')
-        ->condition('cm.moderation_state', ['reject', 'draft'], 'IN')
-        ->condition('cm.workflow', 'service_editorial');
-      $query = $query->execute();
-      $nids = $query->fetchAll();
-      foreach ($nids as $nid) {
-        $this->entityTypeManager->getStorage('node')->deleteRevision($nid->content_entity_revision_id);
-      }
-    }
-
   }
 
 }
